Refactoring moving business Logic to service layer (User)
