+++ app/routes/UserRoute.py
@@ -1,109 +1,34 @@
 #!/usr/bin/env python
 import flask
-from flask import request, jsonify, g, current_app
-from app import db
-from sqlalchemy import or_, and_
+from app.routes import api
+
+from app.routes.validations.errors.ValidationError import ValidationError
+from app.routes.validations.UserCreateInputSchema import UserCreateInputSchema
+from app.shared.HandleRequestValidation import handle_request_validation
+
+from app.shared.Authentication import is_logged
+from flask import current_app, g, request, jsonify
 from flask_httpauth import HTTPBasicAuth
 
-from werkzeug.security import generate_password_hash, check_password_hash
-from app.routes import api
-from app.models.User import User
-from app.models.Company import Company
-from app.routes.validations.UserCreateInputSchema import UserCreateInputSchema
-
-import time
-from datetime import datetime
-from app.shared.Util import format_datetime
-from app.shared.Authentication import is_logged
-
-from app.shared.HandleRequestValidation import handle_request_validation
-from app.routes.validations.errors.ValidationError import ValidationError
+from app.services import UserService
 
 auth = HTTPBasicAuth()
 
 @api.route('/api/users', methods=['POST'])
 def new_user():
-    req_data = request.get_json()
     data_schema = UserCreateInputSchema()
     try:
         handle_request_validation(data_schema)
     except ValidationError as err:
         return jsonify(err.message), 400
 
-    username = req_data['username']
-    email = req_data['email']
-    if 'phone_region' not in req_data:
-        phone_region = '+55'
-    else:
-        phone_region = req_data['phone_region']
-
-    if User.query.filter(or_(User.email==email,User.username==username)).first() is not None:
-        return (jsonify({'message': 'User already exists'}), 400)
-
-    user = User(username=username, 
-                email=email, 
-                company_id=req_data['company_id'], 
-                phone_number=req_data['phone_number'], 
-                phone_region=phone_region, 
-                role_id=req_data['role_id'], 
-                status=1, # ATIVO
-                created_at=datetime.now(),
-                updated_at=datetime.now())
-
-    user.hash_password(req_data['password'])
-    if req_data['role_id'] == 1: # ADMIN
-        company = Company.query.get(req_data['company_id'])
-        company.admin_email = req_data['email']
-        db.session.add(company)
-
-    db.session.add(user)
-    db.session.commit()
-
-    verify_password(user.username, req_data['password'])
-    token = g.user.generate_auth_token(current_app.config['TOKEN_TTL'])
-    response = flask.make_response(jsonify({ 'data': {
-                                        'id': user.id,
-                                        'username': user.username, 
-                                        'company_id': user.company_id,
-                                        'email': user.email,
-                                        'phone_region': user.phone_region,
-                                        'phone_number': user.phone_number,
-                                        'status': user.status,
-                                        'role_id': user.role_id,
-                                        'created_at': format_datetime(user.created_at),
-                                        'updated_at': format_datetime(user.updated_at)}}), 201)
-    response.headers["Content-Type"] = "application/json"
-    setCookie(response,token)
-    return response
-
-def setCookie(response,token):
-    lease = 14 * 24 * 60 * 60  # 14 days in seconds
-    end = time.gmtime(time.time() + lease)
-    expires = time.strftime("%a, %d-%b-%Y %T GMT", end)
-    response.set_cookie('token', token.decode('ascii'), secure=False, domain='.tuamesa.com.br', expires=expires)
+    return UserService.save(request.get_json())
 
 @api.route('/api/users/<int:id>', methods=['GET'])
 def get_user(id):
     if not is_logged():
         return (jsonify({'message': 'Not Authorized' })), 401
-    user = User.query.filter(and_(User.id==id,User.company_id==g.user.company_id)).first()
-    if not user:
-        return (jsonify({'message': 'User not found'}), 404)
-    
-    response = flask.make_response(jsonify({ 'data': {
-                                        'id': user.id,
-                                        'username': user.username, 
-                                        'company_id': user.company_id,
-                                        'email': user.email,
-                                        'phone_region': user.phone_region,
-                                        'phone_number': user.phone_number,
-                                        'status': user.status,
-                                        'role_id': user.role_id,
-                                        'created_at': format_datetime(user.created_at),
-                                        'updated_at': format_datetime(user.updated_at)}}), 200)
-    response.headers["Content-Type"] = "application/json"
-    return response
-
+    return UserService.get(id)
 
 @api.route('/health')
 def get_health():
@@ -114,55 +39,14 @@ def get_health():
 @api.route('/api/users/login', methods=['POST'])
 @auth.login_required
 def get_auth_token():
-    req_data = request.get_json()
-    company = Company.query.get(g.user.company_id)
-
-    if req_data['email'] != company.admin_email:
-        return (jsonify({'message': 'Not Authorized' })), 401
-
-    token = g.user.generate_auth_token(current_app.config['TOKEN_TTL'])
-
-    response = flask.make_response(jsonify({ 'data': {
-                                        'id': g.user.id,
-                                        'username': g.user.username, 
-                                        'company_id': g.user.company_id,
-                                        'email': g.user.email,
-                                        'phone_region': g.user.phone_region,
-                                        'phone_number': g.user.phone_number,
-                                        'status': g.user.status,
-                                        'role_id': g.user.role_id,
-                                        'created_at': format_datetime(g.user.created_at),
-                                        'updated_at': format_datetime(g.user.updated_at)}}), 200)
-    response.headers["Content-Type"] = "application/json"
+  return UserService.get_auth_token(request.get_json())
     
-    setCookie(response,token)
-    return response
-
 @auth.verify_password
 def verify_password(pid, password='password'):
-    # first try to authenticate by token
-    user = User.verify_auth_token(pid)
-    if not user:
-        # try to authenticate with username/password
-        user = User.query.filter_by(username=pid).first()
-        if not user or not user.verify_password(password):
-            return False
-    g.user = user
-    return True
+    return UserService.verify_password(pid, password)
 
 @api.route('/api/users/currentuser')
 def get_current_user():
     if not is_logged():
         return (jsonify({'message': 'Not Authorized' })), 401
-    response = flask.make_response({ 'data': {
-                                        'id': g.user.id,
-                                        'username': g.user.username, 
-                                        'company_id': g.user.company_id,
-                                        'email': g.user.email,
-                                        'phone_region': g.user.phone_region,
-                                        'phone_number': g.user.phone_number,
-                                        'status': g.user.status,
-                                        'role_id': g.user.role_id,
-                                        'created_at': format_datetime(g.user.created_at),
-                                        'updated_at': format_datetime(g.user.updated_at)}}, 200)
-    return response
+    return UserService.get_current_user()
